Section8-GenAI/text-summarization-app/requirements.txt
google-cloud-aiplatform>=1.60.0
google-cloud-storage>=2.14.0
Flask>=3.0.0
gunicorn>=22.0.0
python-docx>=1.1.0
